[feature] Switched trix editor and added autocomplete
I have added trix as a text editor since future changes would be best
suited for this text kind of text editor. Tinymce is complex and hard
to modify, with this in mind I decided to switch to a new text editor,
trix attacted me instantly due to the number of stars and how it is
handling selection and ease of integration. So I have switched to trix.

For the most important feature autocomplete I am using the 'autocomplete'
feature from tribute.js. I am making a ajax request to the varnam api to
receive the suggestions. I am having an issue with selecting the
particular suggestion with enter, I will fix that in the coming commits.
If you want to contribtue on that familarize with the params of the
tribute.js.